Fix MACE disassembler printout for instructions with an address operand.
What the '$' prefix was even meant to mean? It made no sense.
Now, '$' means "current PC", like in the MASM and NASM assemblers, and it is used for branch instructions as the address is PC-relative.
Where the address is absolute, like in LOAD and STORE instructions, the '#' prefix is used instead.

On a side note, it was undoubtely a mistake to make branches take PC-relative addresses and loads/stores absolute addresses, because if they were all relative, the ISA would have been relocation-friendly from the get-go. Not that it matters.